OMAC: Getting 64bit next event from DataTransmissionHandler

+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Samraksh/MAC/OMAC/DataTransmissionHandler.h
@@ -40,7 +40,7 @@ class DataTransmissionHandler: public EventHandler {
 
 public:
 	void Initialize();
-	UINT16 NextEvent(UINT32 currentSlotNum);
+	UINT64 NextEvent(UINT32 currentSlotNum);
 	//UINT64 NextEvent(UINT64 currentTicks);
 	void ExecuteEvent(UINT32 currentSlotNum);
 	UINT8 ExecuteEventDone();
